Rename folder orders to ordersPage and create Order component
