+++ src/pages/index.js
@@ -1,4 +1,4 @@
 export { default as HomePage } from './HomePage/HomePage';
 export { default as InCartPage } from './InCartPage/InCartPage';
 export { default as NoMatchPage } from './NoMatchPage/NoMatchPage';
-export { default as Orders } from './Orders/Orders';
+export { default as OrdersPage } from './OrdersPage/OrdersPage';
